requirements.txt
Flask==0.10.1
Flask-Misaka==0.3.0
Jinja2==2.7.2
MarkupSafe==0.21
Werkzeug==0.9.4
argparse==1.2.1
flake8==2.3.0
itsdangerous==0.24
mccabe==0.3
misaka==1.0.2
pep8==1.6.2
py==1.4.20
pyflakes==0.8.1
pytest==2.5.2
requests==2.5.3
waitress==0.8.9
PyICU==1.9.2
PyYAML==3.11
